Support CoNLL-U files for sequence task
To enable comparisons between different implementations we need to give
a fixed tokenization (as F1-Score and WinPR/Windowdiff depend on token
boundaries). CoNLL-U files enable us to give a segmentation, while
being a widely adopted dataformat.
This commit only prepares for such a comparison as currently the
original tokenization is still lost.

The first step in BERT's tokenization is also whitespace
splitting so our conllu dataset, which is separated at whitespace,
should result in the same final tokenizations.