Likely fix for predicted count not being equal to label count.
The suspected cause is that the last item in the sequence can be reduced
to no token, thereby resulting in `conll_id[x][-1]` not being equal to
`len(conll_labels[x])`.
I was not yet able to isolate a sample causing this bug.
We also now explicitly pick the majority class if a token is not
represented in the BERT tokenized form.